+++ odoo/local-src/delivery_rounds/tests/test_deliveryround_assign.py
@@ -37,15 +37,18 @@ class TestDeliveryRoundAssignMixin(SavepointCase):
 
         cls._add_inventory_qty(cls.p1, 100)
 
-        cls.delivery_template = cls.env['round.template'].create(
+        cls.delivery_template_1 = cls.env['round.template'].create(
+            {'name': 'Unittest delivery template'}
+        )
+        cls.delivery_template_2 = cls.env['round.template'].create(
             {'name': 'Unittest delivery template'}
         )
 
         cls.delivery_round_1 = cls.env['round.instance'].create(
-            {'template_id': cls.delivery_template.id, 'date': '2017-01-01'}
+            {'template_id': cls.delivery_template_1.id, 'date': '2017-01-01'}
         )
         cls.delivery_round_2 = cls.env['round.instance'].create(
-            {'template_id': cls.delivery_template.id, 'date': '2017-01-01'}
+            {'template_id': cls.delivery_template_2.id, 'date': '2017-01-02'}
         )
 
         # pick/ship
@@ -75,28 +78,6 @@ class TestDeliveryRoundAssignMixin(SavepointCase):
         inventory.action_done()
         return inventory
 
-    @classmethod
-    def _prepare_delivery_round(cls):
-        delivery_template = cls.env['round.template'].create(
-            {'name': 'Unittest delivery template'}
-        )
-        delivery_carrier_fixed = cls.env['delivery.carrier'].create(
-            {
-                'name': 'Unittest shipping costs',
-                'delivery_type': 'fixed',
-                'fixed_price': 10.0,
-                'delivery_template_id': delivery_template.id,
-            }
-        )
-        delivery_round = cls.env['round.instance'].create(
-            {
-                'name': 'Unittest delivery round',
-                'template_id': delivery_template.id,
-                'date': '2017-01-01',
-            }
-        )
-        return delivery_carrier_fixed, delivery_round
-
 
 class TestDeliveryRoundAssign(TestDeliveryRoundAssignMixin):
     post_install = True
@@ -107,7 +88,15 @@ class TestDeliveryRoundAssign(TestDeliveryRoundAssignMixin):
         super(TestDeliveryRoundAssign, cls).setUpClass()
 
     def test_deliveryround_carrier(self):
-        delivery_carrier_fixed, delivery_round = self._prepare_delivery_round()
+        # Template 2 = Fixed
+        delivery_carrier_fixed = self.env['delivery.carrier'].create(
+            {
+                'name': 'Unittest shipping costs',
+                'delivery_type': 'fixed',
+                'fixed_price': 10.0,
+                'delivery_template_id': self.delivery_template_2.id,
+            }
+        )
         sale = self.env['sale.order'].create(
             {
                 'partner_id': self.partner.id,
@@ -133,7 +122,9 @@ class TestDeliveryRoundAssign(TestDeliveryRoundAssignMixin):
         sale.action_confirm()
 
         for picking in sale.picking_ids:
-            self.assertEqual(picking.delivery_round_id.id, delivery_round.id)
+            self.assertEqual(
+                picking.delivery_round_id.id, self.delivery_round_2.id
+            )
             self.assertEqual(picking.group_id.carrier_id, sale.carrier_id)
 
     def test_force_deliveryround_partially_available(self):
@@ -230,7 +221,85 @@ class TestDeliveryRoundAssign(TestDeliveryRoundAssignMixin):
         # self.assertEqual(set(pick.pack_operation_ids.mapped('product_id')),
         #                  set([self.p1, self.p2]))
 
-    def test_manual_change_delivery_round(self):
+    def test_manual_change_delivery_round_fixed_to_std(self):
+        delivery_carrier_fixed = self.env['delivery.carrier'].create(
+            {
+                'name': 'Unittest shipping costs',
+                'delivery_type': 'fixed',
+                'fixed_price': 10.0,
+                'delivery_template_id': self.delivery_template_2.id,
+            }
+        )
+        sale = self.env['sale.order'].create(
+            {
+                'partner_id': self.partner.id,
+                'carrier_id': delivery_carrier_fixed.id,
+                'order_line': [
+                    (
+                        0,
+                        0,
+                        {
+                            'name': self.p1.name,
+                            'product_id': self.p1.id,
+                            'product_uom': self.ref(
+                                'product.product_uom_unit'
+                            ),
+                            'product_uom_qty': 3,
+                            'price_unit': 200,
+                        },
+                    )
+                ],
+            }
+        )
+        self.assertFalse(sale.picking_ids)
+        sale.action_confirm()
+
+        for picking in sale.picking_ids:
+            self.assertEqual(
+                picking.delivery_round_id.id, self.delivery_round_2.id
+            )
+            self.assertEqual(picking.group_id.carrier_id, sale.carrier_id)
+
+        ship = sale.picking_ids.filtered(
+            lambda p: p.picking_type_id.code == 'outgoing'
+        )
+        self.env['picking.assign.delivery.round'].with_context(
+            active_ids=ship.ids
+        ).create({'delivery_round_id': self.delivery_round_1.id}).confirm()
+
+        carrier_manual_change = self.env.ref(
+            'delivery_rounds.delivery_carrier_manual_round_change'
+        )
+        for picking in sale.picking_ids:
+            self.assertEqual(
+                picking.delivery_round_id.id, self.delivery_round_1.id
+            )
+            self.assertEqual(
+                picking.group_id.carrier_id, carrier_manual_change
+            )
+
+    def test_manual_change_delivery_round_std_to_fixed(self):
+        self.delivery_round_1.button_resetdraft()
+        self.delivery_round_2.button_resetdraft()
+        self.env['delivery.carrier'].create(
+            {
+                'name': 'Unittest shipping costs',
+                'delivery_type': 'fixed',
+                'fixed_price': 10.0,
+                'delivery_template_id': self.delivery_template_2.id,
+            }
+        )
+        itinerary = self.env['round.itinerary'].create(
+            {
+                'name': 'Itinerary 17C',
+                'code': 'T17C',
+                'sequence': 22,
+                'partner_position_ids': [
+                    (0, 0, {'sequence': 10, 'partner_id': self.partner.id})
+                ],
+            }
+        )
+        self.delivery_round_1.itinerary_ids = itinerary
         sale = self.env['sale.order'].create(
             {
                 'partner_id': self.partner.id,
@@ -254,34 +323,85 @@ class TestDeliveryRoundAssign(TestDeliveryRoundAssignMixin):
         self.assertFalse(sale.picking_ids)
         sale.action_confirm()
 
-        pick = sale.picking_ids.filtered(
-            lambda p: p.picking_type_subcode == 'PICK'
-        )
-
-        self.delivery_round_1._assign_pickings(pick)
         for picking in sale.picking_ids:
             self.assertEqual(
                 picking.delivery_round_id.id, self.delivery_round_1.id
             )
+            self.assertEqual(picking.group_id.carrier_id, sale.carrier_id)
 
-        delivery_round_2 = self.env['round.instance'].create(
-            {'template_id': self.delivery_template.id, 'date': '2017-01-01'}
-        )
         ship = sale.picking_ids.filtered(
             lambda p: p.picking_type_id.code == 'outgoing'
         )
         self.env['picking.assign.delivery.round'].with_context(
             active_ids=ship.ids
-        ).create({'delivery_round_id': delivery_round_2.id}).confirm()
+        ).create({'delivery_round_id': self.delivery_round_2.id}).confirm()
 
         carrier_manual_change = self.env.ref(
             'delivery_rounds.delivery_carrier_manual_round_change'
         )
         for picking in sale.picking_ids:
-            self.assertEqual(picking.delivery_round_id.id, delivery_round_2.id)
+            self.assertEqual(
+                picking.delivery_round_id.id, self.delivery_round_2.id
+            )
             self.assertEqual(
                 picking.group_id.carrier_id, carrier_manual_change
             )
+
+    def test_manual_change_delivery_round_std_to_std(self):
+        self.delivery_round_1.button_resetdraft()
+        self.delivery_round_2.button_resetdraft()
+        itinerary = self.env['round.itinerary'].create(
+            {
+                'name': 'Itinerary 17C',
+                'code': 'T17C',
+                'sequence': 22,
+                'partner_position_ids': [
+                    (0, 0, {'sequence': 10, 'partner_id': self.partner.id})
+                ],
+            }
+        )
+        self.delivery_round_1.itinerary_ids = itinerary
+        sale = self.env['sale.order'].create(
+            {
+                'partner_id': self.partner.id,
+                'order_line': [
+                    (
+                        0,
+                        0,
+                        {
+                            'name': self.p1.name,
+                            'product_id': self.p1.id,
+                            'product_uom': self.ref(
+                                'product.product_uom_unit'
+                            ),
+                            'product_uom_qty': 3,
+                            'price_unit': 200,
+                        },
+                    )
+                ],
+            }
+        )
+        self.assertFalse(sale.picking_ids)
+        sale.action_confirm()
+
+        for picking in sale.picking_ids:
+            self.assertEqual(
+                picking.delivery_round_id.id, self.delivery_round_1.id
+            )
+            self.assertEqual(picking.group_id.carrier_id, sale.carrier_id)
+
+        ship = sale.picking_ids.filtered(
+            lambda p: p.picking_type_id.code == 'outgoing'
+        )
+        self.env['picking.assign.delivery.round'].with_context(
+            active_ids=ship.ids
+        ).create({'delivery_round_id': self.delivery_round_2.id}).confirm()
+
+        for picking in sale.picking_ids:
+            self.assertEqual(
+                picking.delivery_round_id.id, self.delivery_round_2.id
+            )
+            self.assertEqual(picking.group_id.carrier_id, sale.carrier_id)
 
     def test_assign_delivery_round_already_printed(self):
         sale = self.env['sale.order'].create(
@@ -422,8 +542,19 @@ class TestRoundWithCustomerWorkingSchedule(
         delivery_round._assign_pickings(ship1)
         self.assertEqual(ship1.state, 'assigned')
 
-    def test_deliveryround_carrier_schedule(self):
-        delivery_carrier_fixed, delivery_round = self._prepare_delivery_round()
+    def test_deliveryround_carrier_fixed_schedule(self):
+        """
+        Fixed delivery round on carrier. Picking is inserted in the delivery
+        round without taking care of the leave schedule
+        """
+        delivery_carrier_fixed = self.env['delivery.carrier'].create(
+            {
+                'name': 'Unittest shipping costs',
+                'delivery_type': 'fixed',
+                'fixed_price': 10.0,
+                'delivery_template_id': self.delivery_template_2.id,
+            }
+        )
         self.create_schedule(
             {
                 'partner_id': self.partner1.id,
@@ -435,6 +566,44 @@ class TestRoundWithCustomerWorkingSchedule(
             {
                 'partner_id': self.partner1.id,
                 'carrier_id': delivery_carrier_fixed.id,
+                'order_line': [
+                    (
+                        0,
+                        0,
+                        {
+                            'name': self.p1.name,
+                            'product_id': self.p1.id,
+                            'product_uom': self.ref(
+                                'product.product_uom_unit'
+                            ),
+                            'product_uom_qty': 1,
+                            'price_unit': 200,
+                        },
+                    )
+                ],
+            }
+        )
+        self.assertFalse(sale.picking_ids)
+        sale.action_confirm()
+
+        for picking in sale.picking_ids:
+            self.assertEqual(picking.delivery_round_id, self.delivery_round_2)
+
+    def test_deliveryround_carrier_std_schedule(self):
+        """
+        Standard delivery round without specific carrier. Picking is not
+        inserted in the delivery round due to the leave schedule
+        """
+        self.create_schedule(
+            {
+                'partner_id': self.partner1.id,
+                'start_date': '2017-01-01',
+                'end_date': '2017-01-01',
+            }
+        )
+        sale = self.env['sale.order'].create(
+            {
+                'partner_id': self.partner1.id,
                 'order_line': [
                     (
                         0,